refactor: rename action to createAction in internal API
